issue #192 refactor: split shareable components in ScheduleModal
