--- src/components/Common/ScheduleModal/DateAndTime/DateAndTime.styles.js
@@ -1,19 +1,0 @@
-import styled from "styled-components";
-
-// date and time
-export const DateContainerDiv = styled.div`
-	position: relative;
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	margin-bottom: 12px;
-`;
-export const DateDiv = styled.div`
-	display: flex;
-	gap: 15px;
-	width: 45%;
-
-	&:last-child {
-		justify-content: flex-end;
-	}
-`;
